+++ restful_budget_api/library/security.py
@@ -15,7 +15,9 @@ def admin_required(func: F) -> F:
     """decorator requiring server to be running in admin mode"""
 
     @functools.wraps(func)
-    def decorator(*args: Any, **kwargs: Any) -> Union[F, Tuple[Dict[str, str], int]]:
+    def decorator(
+        *args: Any, **kwargs: Any
+    ) -> Union[F, Tuple[Dict[str, str], int]]:
         if current_app.config["IS_ADMIN"]:
             return cast(F, func(*args, **kwargs))
         return ({"error": "server not running in admin mode"}, 403)
@@ -27,17 +29,23 @@ def api_key_required(func: F) -> F:
     """require api_key passed with request json"""
 
     @functools.wraps(func)
-    def decorator(*args: Any, **kwargs: Any) -> Union[F, Tuple[Dict[str, str], int]]:
+    def decorator(
+        *args: Any, **kwargs: Any
+    ) -> Union[F, Tuple[Dict[str, str], int]]:
         if not request.headers.get("Authorization"):
             return (
                 {
-                    "error": "Please provide an API key in the Authorization header of your request"
+                    "error": (
+                        "Please provide an API key in the Authorization "
+                        + "header of your request"
+                    )
                 },
                 400,
             )
         user_key = request.headers.get("Authorization")
         api_keys = [
-            api_key[0] for api_key in db_fetchall(sql="SELECT password FROM users")
+            api_key[0]
+            for api_key in db_fetchall(sql="SELECT password FROM users")
         ]
         if user_key not in api_keys:
             return ({"error": "API key not valid"}, 401)
@@ -50,10 +58,17 @@ def strict_verbiage(func: F) -> F:
     """require function names to match HTTP verbs being requested"""
 
     @functools.wraps(func)
-    def decorator(*args: Any, **kwargs: Any) -> Union[F, Tuple[Dict[str, str], int]]:
+    def decorator(
+        *args: Any, **kwargs: Any
+    ) -> Union[F, Tuple[Dict[str, str], int]]:
         if func.__name__ != request.method.lower():
             return (
-                {"error": f"user {request.method} verb with {func.__name__} method"},
+                {
+                    "error": (
+                        f"user {request.method} verb with {func.__name__} "
+                        + "method"
+                    )
+                },
                 405,
             )
         return cast(F, func(*args, **kwargs))
@@ -67,6 +82,7 @@ def get_user() -> int:
     :param api_key: user's api key
     """
     user_id = db_fetchone(
-        "SELECT id FROM users WHERE password = ?", (request.headers["Authorization"],)
+        "SELECT id FROM users WHERE password = ?",
+        (request.headers["Authorization"],),
     )
     return int(user_id[0])